Move chat files into new folder/ rename some files
